+++ app/src/main/java/com/example/omstugradebook/service/ContactWorkService.java
@@ -26,7 +26,7 @@ public class ContactWorkService {
     private static final String TAG = "Auth Logs";
     private static final String UP_OMGTU = "http://up.omgtu.ru/";
 
-    public List<ContactWork> getContactWork(int userId) {
+    public List<ContactWork> getContactWork(long userId) {
         Document doc = tryToGetConnection(REMOTE_URL);
         if (doc == null) {
             return null;
